Refactored buildFrameInfo function to AnimationCriteria.getFramesInfo()
